Added methods to write to the joystick 'outputs'.
The methods can now write integer value to the IO HID 'outputs' reported by the device.
Note that these are written out one element at a time (instead of all at once) for coding simplicity.
However, it appears that the outputs don't actually do a whole lot (at least on the devices I tested
it with), so the 'Enable Outputs' checkbox in simulink is now always visible.

Also added the library joystick block.